gio: Fix a few -Wsign-conversion warnings

See merge request GNOME/glib!4942

+++ gio/ginetaddressmask.c
@@ -100,7 +100,7 @@ g_inet_address_mask_get_property (GObject    *object,
   switch (prop_id)
     {
     case PROP_FAMILY:
-      g_value_set_enum (value, g_inet_address_get_family (mask->priv->addr));
+      g_value_set_enum (value, (int) g_inet_address_get_family (mask->priv->addr));
       break;
 
     case PROP_ADDRESS:
@@ -312,7 +312,7 @@ g_inet_address_mask_new_from_string (const gchar  *mask_string,
 	  return NULL;
 	}
 
-      address = g_strndup (mask_string, slash - mask_string);
+      address = g_strndup (mask_string, (size_t) (slash - mask_string));
       addr = g_inet_address_new_from_string (address);
       g_free (address);
 
@@ -433,7 +433,7 @@ g_inet_address_mask_matches (GInetAddressMask *mask,
 			     GInetAddress     *address)
 {
   const guint8 *maskbytes, *addrbytes;
-  int nbytes, nbits;
+  size_t nbytes, nbits;
 
   g_return_val_if_fail (G_IS_INET_ADDRESS_MASK (mask), FALSE);
   g_return_val_if_fail (G_IS_INET_ADDRESS (address), FALSE);
